UPSTREAM: edist-test: Fix _Static_assert missing message string
Older GCCs don't support _Static_assert without a message string as the
second argument. AFAICT _Static_assert with two arguments is in C11 but
omitting the message argument is an extension.

The tests appear to be built with the system gcc rather than our
crossgcc so that's probably why this was not cought by CI.

BUG=none
BRANCH=none
TEST=none

Change-Id: I664f9b520ba45f1981cb7a18fc7bc4e9c124d522
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 2aa15872d2468c6244d4c4a1dd0a325a2c9002e9
Original-Change-Id: I41fd0ffc42ded8b6d145c3ec30cc7407a78b9a43
Original-Signed-off-by: Daniel Gröber <dxld@darkboxed.org>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/48151
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2569725
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>