UPSTREAM: soc/mediatek/mt8183: Add DRAM_DMA section
mtk_init_mcu uses DRAM_DMA section as CBFS buffer.
The change "mediatek/mt8183: Remove DRAM_DMA section" is reverted
for using mtk_init_mcu.

On mt8173 and mt8192, this region is used by DMA hardware and is
marked as non-cacheable resource. On mt8183, this region is reserved
as CBFS buffer, so it is not necessary to be marked as non-cacheable
resource.

BUG=none
BRANCH=none
TEST=none

Signed-off-by: Dossym Nurmukhanov <dossym@google.com>
Original-Commit-Id: c221d56478ddb1a364048ea2eb74cb26284eedf8
Original-Change-Id: I7ce9f68883e2787ee7f3c5066f4c47c5ca315633
Original-Signed-off-by: Yidi Lin <yidi.lin@mediatek.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/48232
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Change-Id: I714ed21e0a63591676f39a04ec3b9b611ae1f5f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2585020
Reviewed-by: Dossym Nurmukhanov <dossym@chromium.org>
Tested-by: Dossym Nurmukhanov <dossym@chromium.org>
Commit-Queue: Dossym Nurmukhanov <dossym@chromium.org>