UPSTREAM: mb/google/zork: Set S0IX_SLP_L high in S0, low in S3
This is used as a signal to show the system state.  It hadn't been used
up to this point as we're not currently using S0i3, but the fingerprint
sensor will use it to go into a low power mode, so set it appropriately
on Trembyle.  Dalboz devices don't use the FPMCU, but set there as well
so that the state matches.

BUG=b:174695987
TEST=Verify GPIO state in S0 and S3 with the EC
BRANCH=Zork

Change-Id: Iebb57e311c4bcd03d5aedf1f4ece4c71d3a945e1
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 6a62cc85e61da5282611c2bee5ac3a54580f0113
Original-Signed-off-by: Martin Roth <martinroth@chromium.org>
Original-Change-Id: Ibc725905909830d44f77c2498a26edf6d7a3dc05
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/48255
Original-Reviewed-by: Felix Held <felix-coreboot@felixheld.de>
Original-Reviewed-by: Vincent Palatin <vpalatin@google.com>
Original-Reviewed-by: Rob Barnes <robbarnes@google.com>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2577261
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>