UPSTREAM: azalia: Make `set_bits` function non-static
There's many copies of this function in the tree. Make the copy in
azalia_device.c non-static and rename it to `azalia_set_bits`, then
replace all other copies with it. Since azalia_device.c is only built
when AZALIA_PLUGIN_SUPPORT is selected, select it where necessary.

This has the side-effect of building hda_verb.c from the mainboard
directory. If this patch happens to break audio on a mainboard, it's
because its hda_verb.c was always wrong but wasn't being compiled.

BUG=none
BRANCH=none
TEST=none

Signed-off-by: Dossym Nurmukhanov <dossym@google.com>
Original-Commit-Id: 61dd8365bf442a5c07955ef4b1fd665f8c3aadb4
Original-Change-Id: Iff3520131ec7bc8554612969e3a2fe9cdbc9305e
Original-Signed-off-by: Angel Pons <th3fanbus@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/48346
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Rudolph <siro@das-labor.org>
Change-Id: Iad423251c8761aeee2ff85c4c356a92d532680b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2597685
Reviewed-by: Dossym Nurmukhanov <dossym@chromium.org>
Commit-Queue: Dossym Nurmukhanov <dossym@chromium.org>
Tested-by: Dossym Nurmukhanov <dossym@chromium.org>