UPSTREAM: cbfs: Allow mcache to be found after the first lookup
This patch addresses the same problem as CB:48429, but hopefully this
time correctly. Since the mcache is not guaranteed to be available on
the first CBFS lookup for some special cases, we can no longer treat it
as a one-time fire-and-forget initialization. Instead, we test
cbd->mcache_size to check if the mcache has been initialized yet, and
keep trying on every lookup if we don't find it the first time.

Since the mcache is a hard requirement for TOCTOU safety, also make it
more clear in Kconfig that configurations known to do CBFS accesses
before CBMEM init are incompatbile with that, and make sure we die()
rather than do something unsafe if there's a case that Kconfig didn't
catch.

BUG=none
BRANCH=none
TEST=none

Signed-off-by: Dossym Nurmukhanov <dossym@google.com>
Original-Commit-Id: 34cf0732209e7b9fefc115536719473155e430ea
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Change-Id: I4e01e9a9905f7dcba14eaf05168495201ed5de60
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/48482
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-by: Christian Walter <christian.walter@9elements.com>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Change-Id: I14fc3eae5c60dbdb8fa9d7939553efbaaacf2fe8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2585005
Reviewed-by: Dossym Nurmukhanov <dossym@chromium.org>
Tested-by: Dossym Nurmukhanov <dossym@chromium.org>
Commit-Queue: Dossym Nurmukhanov <dossym@chromium.org>