UPSTREAM: mb/google/volteer: Clean up romstage and ramstage UPDs
Move the manual calls to fw_config_probe() into the devicetree; the
AUDIO probe is trivial, and the TCSS devices (DMA0, iTBT RP0 & RP1) are
already guarded with probe statements in the baseboard devicetree, so
the code in romstage.c was redundant. The variants seem to have their
USB4 probe statements correct as well, so the manual UPD setting in
mainboard.c was also unnecessary.

BUG=none
TEST=abuild google/volteer

Signed-off-by: Joe Tessler <jrt@google.com>
Original-Commit-Id: 2f917e6cee275eda6bd620cb3dbf2ebf17173ed9
Original-Change-Id: I1d067ff3d181b152c784634ff99202bb2b9202f7
Original-Signed-off-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/48512
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Change-Id: Idebc92ea4ce3aedadfaf22fdbdaf7f9adf389702
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2587003
Reviewed-by: Joe Tessler <jrt@chromium.org>
Tested-by: Joe Tessler <jrt@chromium.org>
Commit-Queue: Dossym Nurmukhanov <dossym@chromium.org>