UPSTREAM: configs: Add a weird config for Asus P8Z77-V LX2
This is not meant for actual use, but to build-test several options.
Please do not try to use it on real hardware. Or maybe do try.

The purpose of this config is to build-test the individual options, not
their combination. So, for instance, if it would be hard to keep options
x, y and z build together in the future, this config shouldn't block a
change but should instead be adapted, e.g. split into multiple chunks.

BUG=none
BRANCH=none
TEST=none

Signed-off-by: Dossym Nurmukhanov <dossym@google.com>
Original-Commit-Id: 3ee6d7bf22d4618478ca78f38e2dc648e2f26db3
Original-Change-Id: I80e8fe3982025b61148e7c2b05dd0727d65ee2f4
Original-Signed-off-by: Angel Pons <th3fanbus@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/48546
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Nico Huber <nico.h@gmx.de>
Change-Id: Ic2320b18f7ea582998e25997725f353a4f97e7fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2594029
Reviewed-by: Dossym Nurmukhanov <dossym@chromium.org>
Commit-Queue: Dossym Nurmukhanov <dossym@chromium.org>
Tested-by: Dossym Nurmukhanov <dossym@chromium.org>