UPSTREAM: mb/google/kahlee: move SMI/SCI GPIO setup to ramstage
SMIs and SCIs aren't used before ramstage or the OS, so there should be
no need to already set them up in romstage. Not using this GPIO
configuration functionality allows untangling the GPIO and smi_util code
and only linking smi_util in ramstage in follow-up patches. In romstage
the pins get initialized as inputs with pull-up, so that at least that
part still matches the configuration before this patch.

BUG=b:175386410

Signed-off-by: Dossym Nurmukhanov <dossym@google.com>
Original-Commit-Id: 9ef72ca7db807a7ae2e7d78144773940dd688a78
Original-Change-Id: I733bb91ef60dc66093781a376a2e9837f5209671
Original-Signed-off-by: Felix Held <felix-coreboot@felixheld.de>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/48608
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Martin Roth <martinroth@google.com>
Change-Id: I0749a3731372ba8f845ea8481629d989c318d3ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2590105
Reviewed-by: Dossym Nurmukhanov <dossym@chromium.org>
Tested-by: Dossym Nurmukhanov <dossym@chromium.org>
Commit-Queue: Dossym Nurmukhanov <dossym@chromium.org>