UPSTREAM: mb/ocp/tiogapass: correct "POST complete" pad initial value
On OCP Tioga Pass the pad GPP_B20 is used as output for signalling "POST
complete" to the BMC. According to the schematics and the code in
`ramstage.c`, the signal is active-low. There is an external pull-up
resistor.

To make the signalling work as it should, set the initial output value
to `high`.

BUG=none
BRANCH=none
TEST=none

Change-Id: I652df047d7b02c315b6e1834d994026dd3a82403
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 61469c7a35281b886692db6fc7e1b7114aca4807
Original-Change-Id: I82fbda1caba9163ba3b2e38f494a0cefa27e657f
Original-Signed-off-by: Michael Niewöhner <foss@mniewoehner.de>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/48670
Original-Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2807571
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>