UPSTREAM: util/crossgcc: Add date to the toolchain revision
With the current version method, it's not possible to determine if
a different version is older or newer than the current version without
digging into the repository and finding the dates for the version
numbers.

This change adds the commit date to the start of the toolchain version
which will let us tell at a glance how old or new the toolchain is.

It's not perfect because multiple toolchain commits can go in on the
same day, but adding the time made the string even longer, and really
doesn't help that much.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ib0d3cd5e47fede7fd605adca1c603c2d989dc377
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 7014f8258e6e015fe91d6928266d10ec536e9001
Original-Signed-off-by: Martin Roth <martin@coreboot.org>
Original-Change-Id: I9c6d27667b922dc15e7a6e132e1beff69eed839c
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/48901
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2807692
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>