UPSTREAM: toolchain.inc: Update and fix the test-toolchain target
Due to some change, the test-toolchain was no longer working, and was
always reporting that the toolchain is out of date.

This fixes the failure, and prints both the expected versions of Clang,
GCC, and IASL on failures.

Additional changes fix some indentation issues and skip trying to update
submodules when the test is run.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ic1c0c3b34b8f5a759cc59bf3faf4b03574dac826
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 8be01465286ed24b2a816a8b1cef0b20b55496f6
Original-Signed-off-by: Martin Roth <martin@coreboot.org>
Original-Change-Id: Ia350f279c3fd3533523996327cc6b2304e0bead4
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/48903
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2718800
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>