UPSTREAM: soc/mediatek/mt8192: pmic: update initial setting
We found that the switch frequency of vgpu is at 4~5Mhz with high
current case (> 3.5A) and is at 2.5Mhz with low current case(< 2.8A).
The switch frequency of vgpu should be kept at 2.5Mhz.

The root cause is that phase config of vcore is not disabled, it will
affect the switch frequency of vgpu. Corret the phase setting at
initialization.

BUG=b:172636735
BRANCH=none
TEST=boot asurada correctly

Change-Id: I7ba60651c9a5a7f2c78abe63f65a8879ab6b7bbd
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: cad3f47cda1e804870671ef2031a85c12c593746
Original-Signed-off-by: Hsin-Hsiung Wang <hsin-hsiung.wang@mediatek.com>
Original-Change-Id: I48d3729302de9e3343dce79fe6f5ed045d0296a5
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/49005
Original-Reviewed-by: Yu-Ping Wu <yupingso@google.com>
Original-Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2637286
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>