UPSTREAM: cbfstool: Use flock() when accessing CBFS files
Trying to do multiple operations on the same CBFS image at the same time
likely leads to data corruption. For this reason, add BSD advisory file
locking (flock()) to cbfstool (and ifittool which is using the same file
I/O library), so that only one process will operate on the same file at
the same time and the others will wait in line. This should help resolve
parallel build issues with the INTERMEDIATE target on certain platforms.

Unfortunately, some platforms use the INTERMEDIATE target to do a direct
dd into the CBFS image. This should generally be discouraged and future
platforms should aim to clearly deliminate regions that need to be
written directly by platform scripts with custom FMAP sections, so that
they can be written with `cbfstool write`. For the time being, update
the legacy platforms that do this with explicit calls to the `flock`
utility.

BUG=none
BRANCH=none
TEST=none

Change-Id: I0bf185175048ad1dacc4ef705a7f88915d70561e
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 1153b2ef5c9eb7b12941e08389f1019e18f346e7
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Change-Id: I022468f6957415ae68a7a7e70428ae6f82d23b06
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/49190
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2619706
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>