UPSTREAM: prog_loaders: Remove prog_locate()
This patch rewrites the last few users of prog_locate() to access CBFS
APIs directly and removes the call. This eliminates the double-meaning
of prog_rdev() (referring to both the boot medium where the program is
stored before loading, and the memory area where it is loaded after) and
makes sure that programs are always located and loaded in a single
operation. This makes CBFS verification easier to implement and secure
because it avoids leaking a raw rdev of unverified data outside the CBFS
core code.

BUG=none
BRANCH=none
TEST=none

Change-Id: If3e3a5da88ea32e4febb7bf61953680b91d1f948
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 535846763825f9bc4531b9322b1b61f3973cd6f8
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Change-Id: I7a5525f66e1d5f3a632e8f6f0ed9e116e3cebfcf
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/49337
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2765537
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>